requirements.txt
streamlit>=1.28.0
pandas>=1.5
numpy
matplotlib
scikit-learn
wordcloud
PyPDF2
pdfplumber
python-docx
openpyxl
Pillow
lxml
setuptools
wheel
anthropic>=0.25.0
python-dotenv
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
